Perbarui kodingan

+++ resources/js/Components/TableKeluar.jsx
@@ -8,14 +8,15 @@ import {
   TableRow,
 } from '@/Components/ui/table';
 import { Button } from './ui/button';
-import { Pencil, Trash } from 'lucide-react';
+import { ArrowLeft, Pencil, Trash } from 'lucide-react';
 import { formatDate } from 'date-fns';
 import { router } from '@inertiajs/react';
 import Container from './Container';
+import { AlertDialog, AlertDialogCancel, AlertDialogContent, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from './ui/alert-dialog';
 
 export default function TableKeluar({ stockouts, params, setParams }) {
-    console.log('stockouts ', stockouts)
   const deleteStockOut = (id) => {
+    location.href = `/stock-out?start_date=${params.start_date}&end_date=${params.end_date}`;
     router.delete(`/stock-out/${id}`);
   };
   return (
@@ -46,15 +47,51 @@ export default function TableKeluar({ stockouts, params, setParams }) {
                 </span>
                 <span>Edit</span>
               </Button>
-              <Button
-                onClick={() => deleteStockOut(stock.id)}
-                variant="ghost"
-                className="flex items-center gap-2 text-red-500">
-                <span>
-                  <Trash width={16} height={16} />
-                </span>
-                <span>Hapus</span>
-              </Button>
+              <AlertDialog>
+                  <AlertDialogTrigger asChild>
+                    <Button
+                      variant="ghost"
+                      className="flex items-center gap-2 text-red-500">
+                      <span>
+                        <Trash width={16} height={16} />
+                      </span>
+                      <span>Hapus</span>
+                    </Button>
+                  </AlertDialogTrigger>
+                  <AlertDialogContent>
+                    <AlertDialogHeader>
+                      <AlertDialogTitle>
+                        Apakah anda benar-benar ingin Menghapus stock
+                        {stock.name}?
+                      </AlertDialogTitle>
+                    </AlertDialogHeader>
+                    <AlertDialogFooter>
+                      <AlertDialogCancel>
+                        <Button variant="ghost" className="w-full">
+                          <div className="flex items-center gap-2 ">
+                            <span>
+                              <ArrowLeft width={17} height={17} />
+                            </span>
+                            <span>Kembali</span>
+                          </div>
+                        </Button>
+                      </AlertDialogCancel>
+                      <AlertDialogFooter>
+                        <AlertDialogCancel>
+                          <Button
+                            variant="ghost"
+                            onClick={() => deleteStockOut(stock.id)}
+                            className="flex items-center gap-2 text-red-500">
+                            <span>
+                              <Trash width={16} height={16} />
+                            </span>
+                            <span>Hapus saja</span>
+                          </Button>
+                        </AlertDialogCancel>
+                      </AlertDialogFooter>
+                    </AlertDialogFooter>
+                  </AlertDialogContent>
+                </AlertDialog>
             </TableCell>
           </TableRow>
         ))}
